+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generator configuration by build property
 - Added EmbeddingFileNameType config
 - SourceExpander.Console: Resolve dependency
+- SourceExpander.Console: Show embedded libraries list
 
 ## [4.1.1] - 2022-02-26
 ### Changed